Added functionality:
Login activity now hides the keyboard when the 'log in' button is pressed and shows a snackbar if a wrong username or password has been entered, notifying the user of the error.

List of friends and journeys now clickable and opens their respective fragment when a specific friend/journey is selected.

Made a new class to display selected journey's and/or friend's fragments to minimise repeated code.

Logout button, in the profile fragment, actually logs the user out of the app and displays their username in the username editText field when the login activity is displayed.

Signed-off-by: David <48495648+Daveloper-Dave@users.noreply.github.com>